# backend/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Task Management API",
        "description": "Task Management API Documentation"
    },
    "basePath": "/api",
    "paths": {
        "/tasks": {
            "get": {
                "summary": "Get all tasks",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "summary": "Create a new task",
                "parameters": [
                    {
                        "name": "task",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "put": {
                "summary": "Update a task",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string"
                    },
                    {
                        "name": "task",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "summary": "Delete a task",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "Task": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "completed": {
                    "type": "boolean"
                }
            }
        }
    }
}